disabled deleting objects
